+++ docs/about/history.rst
@@ -17,6 +17,9 @@ Changed
 ^^^^^^^
 .. bullet_list::
 
+   -  Add signed-distance primitives for more accurate clipping and overlapping.
+      See :ref:`clip-surface-section` for more information.
+   -  Add *Render Texture Graphics Format* option to *Clip Surface Sim Settings* to support even more accurate clipping for signed-distance primitives.
    -  Add *Render Texture Graphics Format* option to *Animated Waves Sim Settings* to solve precision issues when using height inputs.
    -  Add XR `SPI` support to *Underwater Renderer*. `[URP]`